cxx11 correctness.

+++ trunk/libsets/Audio1/Rnr/GL/AlListener_GL_Rnr.cxx
@@ -28,8 +28,9 @@ void AlListener_GL_Rnr::Draw(RnrDriver* rd)
     case AlListener::LT_Camera:   tp = rd->GetCamAbsTrans(); break;
     case AlListener::LT_Absolute: tp = &rd->ToGCS();         break;
   }
-  ZTrans& t = *tp;
-  Float_t orient[6] = { t(1,1), t(2,1), t(3,1),  t(1,3), t(2,3), t(3,3) };
+  ZTrans&  t = *tp;
+  Float_t orient[6] = { (float) t(1,1), (float) t(2,1), (float) t(3,1),
+                        (float) t(1,3), (float) t(2,3), (float) t(3,3) };
   alListener3f(AL_POSITION,    t(1,4), t(2,4), t(3,4));
   alListenerfv(AL_ORIENTATION, orient);
 }
